+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.41] - 2020-10-18
+### Added
+- Models.WebClientEntity
+### Changed
+- Views.PageWebClient
+- Views.MainWindow
+
 ## [0.2.38] - 2020-10-17
 ### Added
 - Views.PagePing